frontend/api_postgres/requirements.txt
Django==2.2.13
gunicorn==19.9.0
psycopg2-binary==2.8.3
djangorestframework==3.9.3
markdown
django-filter
jsonschema
jsonpath_ng
